Initial web app
An inital ORT web app that allows creating projects, adding repositories
to project, starting the analyzer, and scanning all found dependencies
with ScanCode. The evaluator and reporter are not yet integrated.

This is just a proof of concept, all components need a lot of work
before they could be used in production.